Prepare a future database file.
This file adds the techniques that the APTs match onto.
The old version of the project doesn't have this feature built into
the database. Since I won't settle for php tomfoolery, I did this in
python. Realistically this shouldn't matter unless MITRE updates the
apt list in the future, in which case you can run the stixx_gen.py
script. There are instructions there for how to do so.